requirements.txt
amqp==2.6.1
asgiref==3.2.10
billiard==3.6.3.0
boto3==1.15.2
botocore==1.18.2
cachetools==4.1.1
celery==4.4.7
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.1.1
decorator==4.4.2
dj-database-url==0.5.0
Django==3.1.1
django-cors-headers==3.5.0
djangorestframework==3.11.1
firebase==3.0.1
gcloud==0.18.3
google-api-core==1.22.2
google-auth==1.21.3
google-cloud-core==1.4.1
google-cloud-storage==1.31.2
google-crc32c==1.0.0
google-resumable-media==1.0.0
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
importlib-metadata==1.7.0
jmespath==0.10.0
jwcrypto==0.8
kombu==4.6.11
moviepy==1.0.3
numpy==1.19.2
oauth2client==4.1.3
opencv-python-headless==4.2.0.32
pafy==0.5.5
Pillow==7.2.0
proglog==0.1.9
protobuf==3.13.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycrypto==2.6.1
python-dateutil==2.8.1
python-decouple==3.3
python-jwt==3.3.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
requests-toolbelt==0.9.1
rsa==4.6
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
sseclient==0.0.27
tqdm==4.50.0
urllib3==1.25.10
uuid==1.30
vine==1.3.0
youtube-dl==2021.03.02
zipp==3.1.0
